+++ editor/editor-core/src/plugins/paste/actions.ts
@@ -1,6 +1,9 @@
+import type { LastContentPasted } from '@atlaskit/editor-plugin-paste';
+
 export enum PastePluginActionTypes {
   START_TRACKING_PASTED_MACRO_POSITIONS = 'START_TRACKING_PASTED_MACRO_POSITIONS',
   STOP_TRACKING_PASTED_MACRO_POSITIONS = 'STOP_TRACKING_PASTED_MACRO_POSITIONS',
+  ON_PASTE = 'ON_PASTE',
 }
 
 export interface StartTrackingPastedMacroPositions {
@@ -10,6 +13,11 @@ export interface StartTrackingPastedMacroPositions {
   };
 }
 
+export interface OnPaste {
+  type: PastePluginActionTypes.ON_PASTE;
+  contentPasted: LastContentPasted;
+}
+
 export interface StopTrackingPastedMacroPositions {
   type: PastePluginActionTypes.STOP_TRACKING_PASTED_MACRO_POSITIONS;
   pastedMacroPositionKeys: string[];
@@ -17,4 +25,5 @@ export interface StopTrackingPastedMacroPositions {
 
 export type PastePluginAction =
   | StartTrackingPastedMacroPositions
-  | StopTrackingPastedMacroPositions;
+  | StopTrackingPastedMacroPositions
+  | OnPaste;