+++ editor/editor-plugin-paste/src/index.ts
@@ -2,4 +2,5 @@ export type {
   PastePlugin,
   PastePluginOptions,
   PastePluginState,
+  LastContentPasted,
 } from './types';